% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosomalSelection.R
\name{autosomalSelection}
\alias{autosomalSelection}
\title{Autosomal Selection}
\usage{
autosomalSelection(AA, Aa, aa, p, it)
}
\arguments{
\item{AA}{Relative fitness of the genotype AA (real between zero and one). Default 1.}

\item{Aa}{Relative fitness of the genotype Aa (real between zero and one).  Default 0.9.}

\item{aa}{Relative fitness of the genotype aa (real between zero and one). Default 0.8}

\item{p}{Initial frequency of A (real between zero and one). Default c(0.10 0.27 0.44 0.61 0.78 0.95).}

\item{it}{Number of generations (integer greater than zero). Default 150.}
}
\value{
A graph of allele frequency, a graph of change of allele frequency as function of allele frequency and, if only one initial frequency, a graph of genotype frequency.
}
\description{
Autosomal Selection
}
\examples{
autosomalSelection()
autosomalSelection(p=0.5)
autosomalSelection(p=c(0.1,0.9))
autosomalSelection(AA=0.5,Aa=1.0,aa=0.5,p=0.5,it=150)
}

