% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftSelection.R
\name{driftSelection}
\alias{driftSelection}
\title{Drift Selection}
\usage{
driftSelection(N, p, it, AA, Aa, aa)
}
\arguments{
\item{N}{Population size (integer greater than zero). Default 100.}

\item{p}{Initial frequency of A (real between zero and one). Default 0.5.}

\item{it}{Number of generations (integer greater than zero). Default 500.}

\item{AA}{Relative fitness of the genotype AA (real between zero and one). Default 0.5.}

\item{Aa}{Relative fitness of the genotype Aa (real between zero and one).  Default 1.0.}

\item{aa}{Relative fitness of the genotype aa (real between zero and one). Default 0.5}
}
\value{
A graph of allele frequency.
}
\description{
Drift Selection
}
\examples{
driftSelection()
driftSelection(p=0.5)
driftSelection(AA=0.5,Aa=1.0,aa=0.5,p=0.5,it=150)
}

