% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evMagic.R
\name{evMagic}
\alias{evMagic}
\title{evMagic}
\usage{
evMagic(magic)
}
\arguments{
\item{magic}{Magic trait (TRUE) or not (FALSE).  Default FALSE}
}
\value{
A multi panel graph of the evolution of naturally selected trait, neutral trait, preference trait, assortative mating metric, intial mating network, and final mating network.
}
\description{
evMagic
}
\examples{
evMagic()
evMagic(magic=TRUE)
evMagic(magic=FALSE)
}

