% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evMagicComplete.R
\name{evMagicComplete}
\alias{evMagicComplete}
\title{evMagicComplete}
\usage{
evMagicComplete(magic, sigma_s, sigma_a, K, b, mig, h)
}
\arguments{
\item{magic}{Magic trait (TRUE) or not (FALSE).  Default FALSE}

\item{sigma_s}{Strength of disruptive selection. Default 0.8}

\item{sigma_a}{Strength of non-random mating. Default 0.1}

\item{K}{Carying capacity. Default 512}

\item{b}{Average nuber of offspring per female. Default 4}

\item{mig}{Migration rate. Default 0.025}

\item{h}{Mutation effect size. Default 0.125}
}
\value{
A multi panel graph of the evolution of naturally selected trait, neutral trait, preference trait, assortative mating metric, intial mating network, and final mating network.
}
\description{
evMagicComplete
}
\examples{
evMagicComplete()
evMagicComplete(magic=TRUE)
evMagicComplete(magic=FALSE)
}

