% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneticDrift.R
\name{geneticDrift}
\alias{geneticDrift}
\title{Genetic Drift}
\usage{
geneticDrift(N, L, p, it, selfing)
}
\arguments{
\item{N}{Population size (integer greater than zero). Default 10}

\item{L}{Number of loci (integer greater than zero).  Default 1.}

\item{p}{Initial frequency of A (real between zero and one). Default p=0.5.}

\item{it}{Number of generations (integer greater than zero). Default 150.}

\item{selfing}{If selfing allowed (TRUE or FALSE). Default FALSE.}
}
\value{
A graph of allele frequency.
}
\description{
Genetic Drift
}
\examples{
geneticDrift()
geneticDrift(p=0.5)
geneticDrift(N=100,L=3,p=0.3,it=500,selfing=TRUE)
}

