% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbreedingCoefficient.R
\name{inbreedingCoefficient}
\alias{inbreedingCoefficient}
\title{Inbreeding Coefficient}
\usage{
inbreedingCoefficient(N, it, initF)
}
\arguments{
\item{N}{Population size (intteger greater than zero). Default 100.}

\item{it}{Number of generations (integer greater than zero). Default 50.}

\item{initF}{Initial inbreeding coefficient (real between zero and one). Default 0.}
}
\value{
A graph of inbreeding coefficient. Ft is the theoretical expectation, Fa is the individual autozygosity, Fp is the population level autozygosity
}
\description{
Inbreeding Coefficient
}
\examples{
inbreedingCoefficient()
inbreedingCoefficient(N=25)
inbreedingCoefficient(N=25,it=500)
inbreedingCoefficient(N=30,initF=0.6)
}

