% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_initVose1991.R
\name{my_initVose1991}
\alias{my_initVose1991}
\title{Calculate coefficient for Vose 1991 algorithm}
\usage{
my_initVose1991(p)
}
\arguments{
\item{p}{A vector of positive real values}
}
\value{
Nothing: internal function
}
\description{
Calculate coefficient for Vose 1991 algorithm
}
\examples{
my_initVose1991(p)
}
\references{
Michael D. Vose (1991) A linear algorithm for generating random numbers with a given distribution. _IEEE Transactions on Software Engineering_, *17*, 972-975.
}

