% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_vose1991.R
\name{my_vose1991}
\alias{my_vose1991}
\title{Return randomdeviates from Vose 1991 algorithm}
\usage{
my_vose1991(N)
}
\arguments{
\item{N}{Number of random deviate returned}
}
\value{
Random deviate
}
\description{
Return randomdeviates from Vose 1991 algorithm
}
\examples{
p<-c(0.1,0.2,0.3,0.4)
my_initVose1991(p)
my_vose1991(4)
}
\references{
Michael D. Vose (1991) A linear algorithm for generating random numbers with a given distribution. _IEEE Transactions on Software Engineering_, *17*, 972-975.
}

