% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationStructure.R
\name{populationStructure}
\alias{populationStructure}
\title{Autosomal Selection}
\usage{
populationStructure(N, p, m, it, Nd, selfing)
}
\arguments{
\item{N}{Population Size in each deme (integer greater than zero). Default 100.}

\item{p}{Initial frequency of A (real between zero and one). Default 0.5.}

\item{m}{Migration rate (real between zero and one). Default 0.01.}

\item{it}{Number of generations (integer greater than zero). Default 300.}

\item{Nd}{Number of demes (integergreater than zero). Default 6.}

\item{selfing}{Is selfing possible (TRUE or FALSE).  Default FALSE.}
}
\value{
A graph of allele frequency and a graph of F-statistics.
}
\description{
Autosomal Selection
}
\examples{
populationStructure()
populationStructure(p=0.5)
populationStructure(N=100,selfing=TRUE)
populationStructure(Nd=10,selfing=FALSE)
}

